// client/src/containers/Payment/enums.ts
/*
 * ENUMERATIONS
 * Enums allow us to organize a collection of related values. Think of them as
 * a class for values, wherein the value can only be a string , or number.
 */

export enum FieldNames {
  EMAIL = 'email',
  FULL_NAME = 'fullName',
  PHONE_NUMBER = 'phoneNumber',
  ADDRESS_CITY = 'address.city',
  ADDRESS_COUNTRY = 'address.country',
  ADDRESS_LINE1 = 'address.line1',
  ADDRESS_LINE2 = 'address.line2',
  ADDRESS_POSTAL_CODE = 'address.postalCode',
  ADDRESS_STATE = 'address.state',
}